Oh god, FINISHED WITH THE AST.  Eveything finds its parrent correctly, we can do a recursive search christ. still need to update the code, but the loop and ifstatment detector runs alot better now.  I put in a bunch of sanity checking so I could use this even on the new bytecode, if I wanted that is.

+++ betteribttest/Instruction.cs
@@ -164,7 +164,7 @@ namespace betteribttest
             {  -4, "noone" },
             {  -5, "global" },
         };
-        public static string lookupInstance(short instance, List<string> InstanceLookup = null)
+        public static string lookupInstance(int instance, List<string> InstanceLookup = null)
         {
             string ret;
             if (instanceLookup.TryGetValue(instance, out ret)) return ret;
